Integrate part of librarian_core auth framework
