+++ broadcast/broadcast.ini
@@ -94,7 +94,7 @@ plugins =
     broadcast.util.squery.database_plugin
     broadcast.util.null_i18n.null_i18n_plugin
     broadcast.util.sessions.session_plugin
-    broadcast.util.auth.user_plugin
+    broadcast.util.auth.plugins.user_plugin
     broadcast.util.static.assets_plugin
 
 routes =
